client/requirements-build.txt
pyinstaller
cython
Pillow
